+++ src/constants/validation.js
@@ -37,8 +37,3 @@ export const sortValidation = (payload) => {
         search: payload,
     };
 };
-
-export const newTodoFormInit = {
-    title: '',
-    description: '',
-};
